Mollify MacOS clang about INT64_FORMAT and jlong
By definition, jlong has to be a 64-bit type, but apparently it is not
the same 64-bit type the INT64_FORMAT specifier expects according
to clang on MacOS.

+++ pljava-so/src/main/c/type/byte_array.c
@@ -54,7 +54,7 @@ static Datum _byte_array_coerceObject(Type self, jobject byteArray)
 			ereport(ERROR, (
 				errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
 				errmsg("cannot accommodate reported Blob length " INT64_FORMAT,
-					length)
+					(int64_t)length)
 			));
 		}
 		byteArray =
